# backend/internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: 影策 HTTP API
  version: "1.0.0"
  description: |
    路径前缀是 `/api`。前后端一起发布，不按路径做 API 版本。

    业务 JSON 信封为 `{ code, data, msg, reason }`。`code === 0` 表示成功；失败时 HTTP status 与
    `code` 都表达失败。`code` 默认等于 HTTP 状态；同一状态下的细分原因使用 `status*100+n`
    （例如 40301 配额、42901 限流）。机器判断请用 `reason`，不要解析 `msg`。

    TypeScript 调用面仍是 `web/src/services/api` 的 `http` 模块，不从本规范生成客户端。
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /admin/channels/{id}/models/batch-reprice:
    post:
      tags: [admin]
      summary: 批量保存系统渠道模型最终销售价
      description: |
        仅管理员可用。接收浏览器计算或手工修改后的最终售价及原始价格版本，不在服务端重算。
        前端按成本/(1−利润率)计算，支持0–6位积分小数四舍五入及重新计算。
        必须提交所选模型全部已启用规格，不改变成本、规格启停或模型启停。
        当前计费方式的价格字段必须完整且不能多余，价格为非负整数微积分，null无效。
        缺少成本/售价配置、重复或失效 ID、越权、超限价格、过期版本或并发版本冲突均整批拒绝。
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [models]
              properties:
                models:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items:
                    type: object
                    required: [modelId, priceVersion, priceTiers]
                    properties:
                      modelId: { type: string, minLength: 1 }
                      priceVersion: { type: integer, format: int64, minimum: 1 }
                      priceTiers:
                        type: array
                        minItems: 1
                        items:
                          type: object
                          required: [id, priceVersion, prices]
                          properties:
                            id: { type: string, minLength: 1 }
                            priceVersion: { type: integer, format: int64, minimum: 1 }
                            prices:
                              type: object
                              additionalProperties: false
                              description: 按次/秒仅unit字段；文本Token为输入、输出、缓存；视频Token仅输出。必须完整提交对应字段。
                              properties:
                                unitPriceMicrocredits: { type: integer, format: int64, minimum: 0, maximum: 9007199254740991 }
                                inputTokenPriceMicrocredits: { type: integer, format: int64, minimum: 0, maximum: 1000000000000 }
                                outputTokenPriceMicrocredits: { type: integer, format: int64, minimum: 0, maximum: 1000000000000 }
                                cachedTokenPriceMicrocredits: { type: integer, format: int64, minimum: 0, maximum: 1000000000000 }
            example: { models: [{ modelId: MODEL001, priceVersion: 4, priceTiers: [{ id: TIER001, priceVersion: 8, prices: { unitPriceMicrocredits: 1370000 } }] }] }
      responses:
        "200":
          description: 更新成功，data.updated 为更新的模型数量
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, enum: [0] }
                  msg: { type: string }
                  data:
                    type: object
                    properties:
                      updated: { type: integer }
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/channels/{id}/models/batch-status:
    post:
      tags: [admin]
      summary: 批量启用或停用系统渠道模型
      description: |
        仅管理员可用。模型 ID 必须属于当前系统渠道，单批 1–100 个模型；
        空 ID、重复 ID、已删除或不属于当前渠道的记录整批拒绝，不会部分更新。
        只修改模型启用状态，并失效模型路由目录缓存。
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [modelIds, enabled]
              properties:
                modelIds:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items: { type: string, minLength: 1 }
                enabled: { type: boolean }
            example: { modelIds: [MODEL001, MODEL002], enabled: false }
      responses:
        "200":
          description: 更新成功，data.updated 为更新的模型数量
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, enum: [0] }
                  msg: { type: string }
                  data:
                    type: object
                    properties:
                      updated: { type: integer }
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/channels/{id}/models/batch-settings:
    post:
      tags: [admin]
      summary: 批量设置系统渠道模型参数
      description: |
        仅管理员可用。批量修改模型启停状态、能力、调用协议、展示名、渠道展示名、描述、Logo 和标签；
        未提交的字段保持原值。模型产品标识、上游模型 ID 和价格不会被修改；能力和调用协议必须同时提交。
        设置能力与协议时，各模型的能力参数将按上游模型 ID 分别重置为协议默认值；价格档不会自动转换，应单独核对。
        模型 ID 必须属于当前系统渠道，单批 1–100 个模型，校验失败时整批拒绝。
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [modelIds]
              properties:
                modelIds:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items: { type: string, minLength: 1 }
                enabled: { type: boolean }
                displayName: { type: string, maxLength: 160 }
                channelLabel: { type: string, maxLength: 80 }
                description: { type: string, maxLength: 500 }
                icon: { type: string, maxLength: 80 }
                tags:
                  type: array
                  maxItems: 5
                  items:
                    type: object
                    required: [text, color]
                    properties:
                      text: { type: string }
                      color: { type: string }
                capability:
                  type: string
                  enum: [text, image, video, audio]
                protocol:
                  type: string
                  minLength: 1
            example: { modelIds: [MODEL001, MODEL002], enabled: true, capability: image, protocol: openai-image }
      responses:
        "200":
          description: 更新成功，data.updated 为更新的模型数量
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, enum: [0] }
                  msg: { type: string }
                  data:
                    type: object
                    properties:
                      updated: { type: integer }
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: open_ai_canvas_session
  schemas:
...
    Envelope:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          description: 0 成功；失败时默认等于 HTTP 状态，或 status*100+n 的细分码
        data: {}
        msg:
          type: string
        reason:
          type: string
          description: 仅失败时出现。稳定机器可读原因。
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            code:
              type: integer
              enum: [400, 401, 403, 404, 409, 429, 500, 502, 503, 504, 40301, 40901, 42901]
            data:
              nullable: true
            reason:
              type: string
              enum:
                - invalid_argument
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - failed_precondition
                - quota_exceeded
                - rate_limited
                - unavailable
                - timeout
                - internal
                - bad_gateway
                - upstream_dns_failed
                - model_capability_not_supported
                - model_price_not_configured
                - model_route_unavailable
                - provider_request_failed
                - model_catalog_mismatch
                - invalid_model_selection
  responses:
...
    Error:
      description: 业务失败信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"